+++ util.c
@@ -559,7 +559,7 @@ static void glob_expand_one(char *s, char **argv, int *argc, int maxargs)
 		globfree(&globbuf);
 		return;
 	}
-	for (i=0; i<(maxargs - (*argc)) && i<globbuf.gl_pathc;i++) {
+	for (i=0; i<(maxargs - (*argc)) && i < (int) globbuf.gl_pathc;i++) {
 		if (i == 0) free(argv[*argc]);
 		argv[(*argc) + i] = strdup(globbuf.gl_pathv[i]);
 		if (!argv[(*argc) + i]) out_of_memory("glob_expand");
